# cpu: risc-v: pooling: further optimize the rv64 maxpool